In seash_helper.py: enabled Affix for sockettimeout; cleaned up imports; improved code style.
The biggest problem here is that we require a library that is also
used by the Affix framework, `sockettimeout.r2py`. If Affixes are
enabled, they will import that library, and its functions will
bind to the original Repy network calls. Yet, when we import it,
we want its functions to use the Affix'ed versions of the calls!
The hack I came up with is that I use one of `sockettimeout`'s
wrapper functions on a socket that I get from an overridden,
Affix-enabled `openconnection` call (instead of
`sockettimeout.timeout_openconnection` which will use the original
Repy `openconnection`).

Cleanups relate to using `dy_import_module` rather than the
`...symbols` version so we import libraries only once, and
associated adding of module names to function calls.
Also cleaned up a bit of sub-indentation.